refactor: share chat protocol sanitizers
